+++ example/lib/main.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_html/flutter_html.dart';
 import 'package:new_version_plus/new_version_plus.dart';
 
 void main() {
@@ -30,8 +31,9 @@ class _MyHomePageState extends State<MyHomePage> {
     super.initState();
 
     // Instantiate NewVersion manager object (Using GCP Console app as example)
-    final newVersion = NewVersionPlus(iOSId: 'com.disney.disneyplus', androidId: 'com.disney.disneyplus', androidPlayStoreCountry: "es_ES" //support country code
-        );
+    final newVersion = NewVersionPlus(
+      iOSId: 'com.disney.disneyplus', androidId: 'com.whatsapp', androidPlayStoreCountry: "es_ES", androidHtmlReleaseNotes: true, //support country code
+    );
 
     // You can let the plugin handle fetching the status and showing a dialog,
     // or you can fetch the status and display your own dialog, or no dialog.
@@ -92,7 +94,7 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
       body: SafeArea(
         child: Center(
-          child: Text(release),
+          child: Html(data: release),
         ),
       ),
     );